refactor: extract group QR code modal into GroupQrCodeModal component
Extract the inline QR code modal and all related logic (~550 lines) from
ChatHeader.vue into a dedicated GroupQrCodeModal.vue component to improve
maintainability and follow the existing component extraction pattern.
